BokehRenderer: mv rendering() outside class. ColoredImage: add get_line_and_color_key. Let {x,y}{0,1} asfloat.

+++ TraceDisplay/shape/line/sched_rq_size_change.py
@@ -2,8 +2,8 @@ def sched_rq_size_change(df):
     import numpy as np
     df = df['sched_rq_size_change']
     size = np.array(df['size'])
-    timestamp = np.array(df.index)
-    cpu = np.array(df['cpu'])
+    timestamp = np.array(df.index, dtype=float)
+    cpu = np.array(df['cpu'], dtype=float)
     return {
         'x0' : timestamp,
         'x1' : timestamp,
