+++ TraceDisplay/shape/line/sched_rq_size_interval.py
@@ -2,9 +2,9 @@ def sched_rq_size_interval(df):
     import numpy as np
     df = df['sched_rq_size_change']
     size = np.array(df['size'])
-    timestamp = np.array(df.index)
-    next_timestamp_on_same_cpu = np.array(df.index)
-    cpu = np.array(df['cpu'])
+    timestamp = np.array(df.index, dtype=float)
+    next_timestamp_on_same_cpu = np.array(df.index, dtype=float)
+    cpu = np.array(df['cpu'], dtype=float)
     idx = np.arange(len(next_timestamp_on_same_cpu))
     # TODO: in parallel
     for i in np.unique(cpu):
